优化弹射物穿透

+++ src/main/java/com/atsuishio/superbwarfare/entity/projectile/SmallCannonShellEntity.java
@@ -17,6 +17,7 @@ import net.minecraft.network.protocol.game.ClientGamePacketListener;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.sounds.SoundSource;
+import net.minecraft.util.Mth;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
@@ -142,7 +143,10 @@ public class SmallCannonShellEntity extends FastThrowableProjectile implements G
             float hardness = this.level().getBlockState(resultPos).getBlock().defaultDestroyTime();
             if (hardness != -1) {
                 if (ExplosionConfig.EXPLOSION_DESTROY.get() && this.blockInteraction == null) {
-                    this.level().destroyBlock(resultPos, true);
+                    boolean destroy = Math.random() < Mth.clamp(1 - (hardness / 50), 0.1 , 1);
+                    if (destroy) {
+                        this.level().destroyBlock(resultPos, true);
+                    }
                 }
             }
         }
